+++ src/translation/fr-FR/index.ts
@@ -5,11 +5,13 @@ import {KitCollapseLocale} from './components/DataDisplay/Collapse';
 import {KitEmptyLocale} from './components/DataDisplay/Empty';
 import {KitErrorLocale} from './components/DataDisplay/Error';
 import {KitGeneralLocale} from './general';
+import {KitAlertLocale} from '@translation/en-US/components/Alert';
 
 export const frFR: IKitLocale = {
     locale: 'frFR',
     General: KitGeneralLocale,
     Collapse: KitCollapseLocale,
+    Alert: KitAlertLocale,
     Empty: KitEmptyLocale,
     Error: KitErrorLocale,
     Image: KitImageLocale,