+++ src/translation/types/components/Feedback/Alert/index.d.ts
@@ -1,0 +1,3 @@
+export interface IKitAlertLocale {
+    showDetails?: string;
+}