+++ src/translation/types/index.d.ts
@@ -5,6 +5,7 @@ import {IKitCollapseLocale} from './components/DataDisplay/Collapse';
 import {IKitEmptyLocale} from './components/DataDisplay/Empty';
 import {IKitErrorLocale} from './components/DataDisplay/Error';
 import {IKitGeneralLocale} from './general';
+import {IKitAlertLocale} from './components/Feedback/Alert';
 
 type LanguageSupported = 'frFR' | 'enUS';
 
@@ -12,6 +13,7 @@ export interface IKitLocale extends Locale {
     locale?: LanguageSupported;
     General?: IKitGeneralLocale;
     Collapse?: IKitCollapseLocale;
+    Alert?: IKitAlertLocale;
     Empty?: IKitEmptyLocale;
     Error?: IKitErrorLocale;
     Image?: IKitImageLocale;