+++ stories/02-Design System/06-Feedback/Alert/examples/icon.tsx
@@ -2,14 +2,6 @@ import {KitAlert} from '@kit/index';
 
 const App = () => (
     <>
-        <KitAlert message="Success Tips" type="success" showIcon />
-        <br />
-        <KitAlert message="Informational Notes" type="info" showIcon />
-        <br />
-        <KitAlert message="Warning" type="warning" showIcon closable />
-        <br />
-        <KitAlert message="Error" type="error" showIcon />
-        <br />
         <KitAlert
             message="Success Tips"
             description="Detailed description and advice about successful copywriting."
@@ -24,13 +16,7 @@ const App = () => (
             showIcon
         />
         <br />
-        <KitAlert
-            message="Warning"
-            description="This is a warning notice about copywriting."
-            type="warning"
-            showIcon
-            closable
-        />
+        <KitAlert message="Warning" description="This is a warning notice about copywriting." type="warning" showIcon />
         <br />
         <KitAlert message="Error" description="This is an error message about copywriting." type="error" showIcon />
     </>