+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [X.X.X](https://www.npmjs.com/package/aristid-ds/v/X.X.X) (202X-XX-XX)
 
 ### ✨ Features
+-   **⚠️ Breaking change - Alert:**
+    - Update Alert component to make it toastable. Does not rely on Ant anymore. Props remains the same except `afterClose` and `banner` which are removed. (@evoiron)
+
 -   **⚠️ Breaking change - Grid:**
     - Update KitRow and KitCol to use antd API (@Delmotte-Vincent)